Refactor comment configuration for Entity Framework Core models
