+++ src/EntityFrameworkCore/BitzArt.XDoc.EntityFrameworkCore/Extensions/ModelBuilderExtensions.cs
@@ -1,6 +1,5 @@
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata;
 
 namespace BitzArt.XDoc;
 
@@ -10,8 +9,6 @@ namespace BitzArt.XDoc;
 [PublicAPI]
 public static class ModelBuilderExtensions
 {
-    private static List<string> _processedColumns = new List<string>();
-    
     /// <summary>
     /// Configures XML documentation comments for Entity Framework Core entities and their properties.
     /// </summary>
@@ -19,100 +16,9 @@ public static class ModelBuilderExtensions
     /// <param name="xDoc"></param>
     public static ModelBuilder ConfigureComments(this ModelBuilder modelBuilder, XDoc xDoc)
     {
-        var entityTypes = modelBuilder.Model.GetEntityTypes();
-
-        foreach (var entityType in entityTypes)
-        {
-            ConfigureEntity(modelBuilder, xDoc, entityType);
-        }
+        var commentsConfigurator = new CommentsConfigurator(xDoc);
+        commentsConfigurator.SetComments(modelBuilder);
 
         return modelBuilder;
     }
-
-    private static void ConfigureEntity(ModelBuilder modelBuilder, XDoc xDoc, IMutableEntityType entityType)
-    {
-        ConfigureEntityTypeComment(xDoc, entityType);
-
-        var properties = entityType.GetProperties();
-
-        foreach (var property in properties)
-        {
-            ConfigureEntityPropertyComment(modelBuilder, xDoc, entityType, property);
-        }
-    }
-
-    private static void ConfigureEntityTypeComment(XDoc xDoc, IMutableEntityType entityType)
-    {
-        var typeDocumentation = xDoc.Get(entityType.ClrType);
-
-        if (typeDocumentation is null) 
-        {
-            // No own xml-documentation
-            return;
-        }
-        
-        var isOwned = entityType.IsOwned();
-        var tableName = entityType.GetTableName();
-
-        if (isOwned || tableName is null)
-        {
-            // For owned entities, we don't set the comment on the entity itself
-            // But we will set the comment on the properties
-            
-            return;
-        }
-        
-        if (entityType.BaseType != null && entityType.BaseType.GetTableName() == tableName)
-        {
-            // Base type already defined comment
-            return;
-        }
-        
-        var entityComment = typeDocumentation.ToPlainText();
-            
-        entityType.SetComment(entityComment);
-    }
-    
-    private static void ConfigureEntityPropertyComment(
-        ModelBuilder modelBuilder,
-        XDoc xDoc,
-        IMutableEntityType entityType,
-        IMutableProperty property)
-    {
-        var columnName = property.GetColumnName();
-        var tableName = entityType.GetTableName();
-
-        if (_processedColumns.Contains($"{tableName}.{columnName}"))
-        {
-            return;
-        }
-        
-        _processedColumns.Add($"{tableName}.{columnName}");
-        
-        var isShadowProperty = property.IsShadowProperty();
-
-        if (isShadowProperty)
-        {
-            return;
-        }
-
-        var propertyInfo = entityType.ClrType.GetProperty(property.Name);
-
-        if (propertyInfo is null)
-        {
-            return;
-        }
-
-        var propertyDocumentation = xDoc.Get(propertyInfo);
-
-        if (propertyDocumentation is null)
-        {
-            return;
-        }
-
-        var propertyComment = propertyDocumentation.ToPlainText();
-
-        property.SetComment(propertyComment);
-    }
-    
 }
